requirements.txt
streamlit
openai
pydub
python-docx
reportlab

